Messing around with model loading- read desc.
Added code for loading of additional tank model, had to make a change to
m_Model in Transform.cpp (read commit changes for details) to allow this
to work properly.
Also added a new folder called 'Models' to contain all models and
textures to be used - have to copy and paste required models and
textures from this folder into normal 'textures' and 'models' folders in
'assets' folder for use in program.